basis for exclude arms

+++ src/test/resources/schema.sql
@@ -134,7 +134,7 @@ CREATE TABLE Model (
 -- changeset reidd:6
 CREATE TABLE ArmExclusion (
   id SERIAL NOT NULL,
-  trialverseId INT NOT NULL,
+  trialverseId BIGINT NOT NULL,
   analysisId INT NOT NULL,
   PRIMARY KEY (id),
   FOREIGN KEY(analysisId) REFERENCES NetworkMetaAnalysis(id)
